enum class HostFunctionError
